+++ test/test_geod.py
@@ -4,7 +4,7 @@ import os
 import pickle
 import shutil
 import tempfile
-from contextlib import contextmanager
+from contextlib import contextmanager, nullcontext
 from itertools import permutations
 
 import numpy as np
@@ -12,7 +12,7 @@ import pytest
 from numpy.testing import assert_almost_equal
 
 from pyproj import Geod
-from pyproj.geod import GeodIntermediateFlag
+from pyproj.geod import GeodIntermediateFlag, reverse_azimuth
 
 try:
     from shapely.geometry import (
@@ -70,17 +70,27 @@ def test_geod_inverse_transform():
     true_az12 = -66.5305947876623
     true_az21 = 75.65363415556968
     print("from pyproj.Geod.inv:")
-    az12, az21, dist = gg.inv(lon1pt, lat1pt, lon2pt, lat2pt)
-    assert_almost_equal(
-        (az12, az21, dist), (true_az12, true_az21, 4164192.708), decimal=3
-    )
+    for return_back_azimuth in (True, False):
+        az12, az21, dist = gg.inv(
+            lon1pt, lat1pt, lon2pt, lat2pt, return_back_azimuth=return_back_azimuth
+        )
+        if not return_back_azimuth:
+            az21 = reverse_azimuth(az21)
+        assert_almost_equal(
+            (az12, az21, dist), (true_az12, true_az21, 4164192.708), decimal=3
+        )
 
     print("forward transform")
     print("from proj.4 geod:")
-    endlon, endlat, backaz = gg.fwd(lon1pt, lat1pt, az12, dist)
-    assert_almost_equal(
-        (endlon, endlat, backaz), (lon2pt, lat2pt, true_az21), decimal=3
-    )
+    for return_back_azimuth in (True, False):
+        endlon, endlat, backaz = gg.fwd(
+            lon1pt, lat1pt, az12, dist, return_back_azimuth=return_back_azimuth
+        )
+        if not return_back_azimuth:
+            backaz = reverse_azimuth(backaz)
+        assert_almost_equal(
+            (endlon, endlat, backaz), (lon2pt, lat2pt, true_az21), decimal=3
+        )
 
     inc_exc = ["excluding", "including"]
     res_az12_az21_dists_all = [
@@ -150,32 +160,40 @@ def test_geod_inverse_transform():
             ).transpose()
 
     del_s = dist / (point_count - 1)
-    lons_a = np.empty(point_count)
-    lats_a = np.empty(point_count)
-    azis_a = np.empty(point_count)
+    for return_back_azimuth in [None, True, False]:
+        print(f"test inv_intermediate (by npts) with azi output {return_back_azimuth=}")
+        lons_a = np.empty(point_count)
+        lats_a = np.empty(point_count)
+        azis_a = np.empty(point_count)
 
-    print("test inv_intermediate (by npts) with azi output")
-    res = gg.inv_intermediate(
-        out_lons=lons_a,
-        out_lats=lats_a,
-        out_azis=azis_a,
-        lon1=lon1pt,
-        lat1=lat1pt,
-        lon2=lon2pt,
-        lat2=lat2pt,
-        npts=point_count,
-        initial_idx=0,
-        terminus_idx=0,
-    )
-    assert res.npts == point_count
-    assert_almost_equal(res.del_s, del_s)
-    assert_almost_equal(res.dist, dist)
-    assert_almost_equal(res.lons, lons)
-    assert_almost_equal(res.lats, lats)
-    assert_almost_equal(res.azis[:-1], azis12[1:])
-    assert res.lons is lons_a
-    assert res.lats is lats_a
-    assert res.azis is azis_a
+        with pytest.warns(
+            UserWarning
+        ) if return_back_azimuth is None else nullcontext():
+            res = gg.inv_intermediate(
+                out_lons=lons_a,
+                out_lats=lats_a,
+                out_azis=azis_a,
+                lon1=lon1pt,
+                lat1=lat1pt,
+                lon2=lon2pt,
+                lat2=lat2pt,
+                npts=point_count,
+                initial_idx=0,
+                terminus_idx=0,
+                return_back_azimuth=return_back_azimuth,
+            )
+        assert res.npts == point_count
+        assert_almost_equal(res.del_s, del_s)
+        assert_almost_equal(res.dist, dist)
+        assert_almost_equal(res.lons, lons)
+        assert_almost_equal(res.lats, lats)
+        out_azis = res.azis[:-1]
+        if return_back_azimuth in [True, None]:
+            out_azis = reverse_azimuth(out_azis)
+        assert_almost_equal(out_azis, azis12[1:])
+        assert res.lons is lons_a
+        assert res.lats is lats_a
+        assert res.azis is azis_a
 
     for flags in (GeodIntermediateFlag.AZIS_DISCARD, GeodIntermediateFlag.AZIS_KEEP):
         print("test inv_intermediate (by npts) without azi output, no buffers")
@@ -188,6 +206,7 @@ def test_geod_inverse_transform():
             initial_idx=0,
             terminus_idx=0,
             flags=flags,
+            return_back_azimuth=False,
         )
         assert res.npts == point_count
         assert_almost_equal(res.del_s, del_s)
@@ -216,41 +235,53 @@ def test_geod_inverse_transform():
             initial_idx=0,
             terminus_idx=0,
             flags=flags,
+            return_back_azimuth=False,
         )
         assert res.npts == point_count
-        assert_almost_equal(res.del_s, del_s)
-        assert_almost_equal(res.dist, dist)
-        assert_almost_equal(res.lons, lons_a)
-        assert_almost_equal(res.lats, lats_a)
         assert res.lons is lons_b
         assert res.lats is lats_b
         if flags == GeodIntermediateFlag.AZIS_DISCARD:
             assert res.azis is None
         else:
             assert_almost_equal(res.azis, azis_a)
+        assert_almost_equal(res.del_s, del_s)
+        assert_almost_equal(res.dist, dist)
+        assert_almost_equal(res.lons, lons_a)
+        assert_almost_equal(res.lats, lats_a)
 
-    print("test fwd_intermediate")
-    res = gg.fwd_intermediate(
-        out_lons=lons_b,
-        out_lats=lats_b,
-        out_azis=azis_b,
-        lon1=lon1pt,
-        lat1=lat1pt,
-        azi1=true_az12,
-        npts=point_count,
-        del_s=del_s,
-        initial_idx=0,
-        terminus_idx=0,
-    )
-    assert res.npts == point_count
-    assert_almost_equal(res.del_s, del_s)
-    assert_almost_equal(res.dist, dist)
-    assert_almost_equal(res.lons, lons_a)
-    assert_almost_equal(res.lats, lats_a)
-    assert_almost_equal(res.azis, azis_a)
-    assert res.lons is lons_b
-    assert res.lats is lats_b
-    assert res.azis is azis_b
+    for return_back_azimuth in [False, True, None]:
+        print(f"test fwd_intermediate with {return_back_azimuth=}")
+        lons_b = np.empty(point_count)
+        lats_b = np.empty(point_count)
+        azis_b = np.empty(point_count)
+
+        with pytest.warns(
+            UserWarning
+        ) if return_back_azimuth is None else nullcontext():
+            res = gg.fwd_intermediate(
+                out_lons=lons_b,
+                out_lats=lats_b,
+                out_azis=azis_b,
+                lon1=lon1pt,
+                lat1=lat1pt,
+                azi1=true_az12,
+                npts=point_count,
+                del_s=del_s,
+                initial_idx=0,
+                terminus_idx=0,
+                return_back_azimuth=return_back_azimuth,
+            )
+        assert res.npts == point_count
+        assert res.lons is lons_b
+        assert res.lats is lats_b
+        assert res.azis is azis_b
+        assert_almost_equal(res.del_s, del_s)
+        assert_almost_equal(res.dist, dist)
+        assert_almost_equal(res.lons, lons_a)
+        assert_almost_equal(res.lats, lats_a)
+        if return_back_azimuth in [True, None]:
+            azis_b = reverse_azimuth(azis_b)
+        assert_almost_equal(azis_b, azis_a)
 
     print("test inv_intermediate (by del_s)")
     for del_s_fact, flags in (
@@ -270,6 +301,7 @@ def test_geod_inverse_transform():
             initial_idx=0,
             terminus_idx=0,
             flags=flags,
+            return_back_azimuth=False,
         )
         assert res.npts == point_count
         assert_almost_equal(res.del_s, del_s)
@@ -690,3 +722,24 @@ def test_geod__build_kwargs(kwarg):
     assert_almost_equal(gg.b, gg2.b)
     assert_almost_equal(gg.f, gg2.f)
     assert_almost_equal(gg.es, gg2.es)
+
+
+@pytest.mark.parametrize("radians", [False, True])
+def test_geod__reverse_azimuth(radians):
+    f = math.pi / 180 if radians else 1
+    xy = np.array(
+        [
+            [0, 0 + 180],
+            [180, 180 - 180],
+            [-180, -180 + 180],
+            [10, 10 - 180],
+            [20, 20 - 180],
+            [-10, -10 + 180],
+        ]
+    )
+    for x, y in xy:
+        assert_almost_equal(reverse_azimuth(x * f, radians=radians), y * f)
+
+    xx = xy.T[0]
+    yy = xy.T[1]
+    assert_almost_equal(reverse_azimuth(xx * f, radians=radians), yy * f)